fix: consolidate schema files and update test imports
