--- platform/flowglad-next/src/lib/authSchema.ts
@@ -1,40 +1,0 @@
-import { z } from 'zod'
-
-export const PASSWORD_MIN_LENGTH = 8
-
-export const passwordSchema = z
-  .string()
-  .min(
-    PASSWORD_MIN_LENGTH,
-    `Password must be at least ${PASSWORD_MIN_LENGTH} characters`
-  )
-
-export const signupSchema = z
-  .object({
-    firstName: z.string().min(1, 'First name is required'),
-    lastName: z.string().min(1, 'Last name is required'),
-    email: z.email('Invalid email address'),
-    password: passwordSchema,
-    passwordConfirmation: z.string(),
-  })
-  .refine((data) => data.password === data.passwordConfirmation, {
-    path: ['passwordConfirmation'],
-    message: 'Passwords do not match',
-  })
-
-export const signInSchema = z.object({
-  email: z.email({ message: 'Please enter a valid email' }),
-  password: z
-    .string()
-    .min(1, { message: 'Please enter your password' }),
-})
-
-export const newPasswordSchema = z
-  .object({
-    password: passwordSchema,
-    passwordConfirmation: z.string(),
-  })
-  .refine((data) => data.password === data.passwordConfirmation, {
-    path: ['passwordConfirmation'],
-    message: 'Passwords do not match',
-  })
