+++ platform/flowglad-next/src/lib/schemas.ts
@@ -2,7 +2,7 @@ import { z } from 'zod'
 
 export const PASSWORD_MIN_LENGTH = 8
 
-const passwordSchema = z
+export const passwordSchema = z
   .string()
   .min(
     PASSWORD_MIN_LENGTH,
